ap-2482: Move shared submissions processing code to a concern
